Comments component created 1. Created a folder for Comments components and moved files for comments into folder 2. Installed React read more less 3. User comments from API are now showing
There are no errors in the console at this time.
