+++ package.json
@@ -21,6 +21,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
+    "react-read-more-less": "^0.1.6",
     "react-router-dom": "^6.11.0",
     "react-scripts": "5.0.1",
     "sass": "^1.66.1",
