+++ src/components/businessPage/BusinessPage.js
@@ -8,9 +8,9 @@ import Tab from "react-bootstrap/Tab";
 import Tabs from "react-bootstrap/Tabs";
 // import ShowMap from "../ShowMap";
 
-import Comments from "../Comments";
-import Comment from "../Comment";
-import CommentForm from "../CommentForm";
+import Comments from "../comments/Comments";
+import Comment from "../comments/Comment";
+import CommentForm from "../comments/CommentForm";
 import { StarRatingAndReviews } from "../StarRating";
 import BusinessHours from "../businessHours/BusinessHours";
 
@@ -221,14 +221,21 @@ const Show = ({ setFavs, favs, currentUser, findBusinessByPlaceId }) => {
             <h4>Ratings and reviews</h4>
             {businessDataFromAPI && (
               <span>
-                <p className="Rating">
-                {businessDataFromAPI.rating}</p>
-              <StarRatingAndReviews
-                rating={businessDataFromAPI.rating}
-                reviews={businessReviews}
+                <p className="Rating">{businessDataFromAPI.rating}</p>
+                <StarRatingAndReviews
+                  rating={businessDataFromAPI.rating}
+                  reviews={businessReviews}
+                />
+              </span>
+            )}
+            {businessDataFromAPI.reviews?.slice(0, 1).map((comment, index) => (
+              <Comment
+                key={index}
+                comment={comment}
+                handleDelete={handleDelete}
+                handleSubmit={handleEdit}
               />
-            </span>
-          )}
+            ))}
           </div>
           <div className="BusinessPage__Details__Expanded__Details">
             <h4>Details</h4>
@@ -241,27 +248,29 @@ const Show = ({ setFavs, favs, currentUser, findBusinessByPlaceId }) => {
             <h5>Location and contact</h5>
             <p>Map</p>
             {!is_store ? (
-            <a href={website ? website : "N/A"} target="*">
-              Online Only
-            </a>
-          ) : (
+              <a href={website ? website : "N/A"} target="*">
+                Online Only
+              </a>
+            ) : (
+              <p>
+                <i class="fa-solid fa-location-dot"></i>{" "}
+                <a href={`http://maps.google.com/?q=${name}`} target="*">
+                  {businessDataFromAPI.formatted_address || address}
+                </a>
+              </p>
+            )}
             <p>
-              <i class="fa-solid fa-location-dot"></i>{" "}
-              <a href={`http://maps.google.com/?q=${name}`} target="*">
-                {businessDataFromAPI.formatted_address || address}
+              <i class="fa-solid fa-laptop"></i>{" "}
+              <a href={website ? website : "N/A"} target="*">
+                Website
               </a>
             </p>
-            )}
-                      <p>
-            <i class="fa-solid fa-laptop"></i>{" "}
-            <a href={website ? website : "N/A"} target="*">
-              Website
-            </a>
-            </p>
             <a href="#">
-            <i class="fa-solid fa-phone"></i>{" "}
-            {businessDataFromAPI.formatted_phone_number || contact_num || "N/A"}
-          </a>
+              <i class="fa-solid fa-phone"></i>{" "}
+              {businessDataFromAPI.formatted_phone_number ||
+                contact_num ||
+                "N/A"}
+            </a>
           </div>
         </div>
