+++ src/components/comments/Comment.scss
@@ -1,0 +1,6 @@
+.Comment {
+  &__Commenter {
+    list-style: none;
+    font-size: 14px;
+  }
+}
